ref: revert back some changes to match the staging
Remove the pending-approvals tab on MeSidebar component and to me page, and put it on dept page instead.

Revert version of middleware to proxy (proxy is the newer version of middleware, and middleware is deprecated)

Revert version of other stuff, such as signup, signin actions, SignUp, SignIn components.

Use allowedDevOrigins instead of allowedOrigins, for development purposes and give easy access for codespaces. (This supposed to be for development only, other origins are not allowed on production)
