+++ next.config.mjs
@@ -1,10 +1,10 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  experimental: {
-    serverActions: {
-      allowedOrigins: ['localhost:3000', 'supreme-fortnight-69xj6rqx554ghrv64-3000.app.github.dev', 'supreme-fortnight-69xj6rqx554ghrv64-3001.app.github.dev'],
-    },
-  },
+  allowedDevOrigins: [
+    '*.app.github.dev',
+    'localhost:3000',
+    'localhost:3001'
+  ],
 };
 
 export default nextConfig;
