+++ src/app/actions/auth/signin.js
@@ -4,10 +4,7 @@ import { createClient } from "lib/supabase/server";
 import { redirect } from "next/navigation";
 import { getRedirectPath } from "lib/supabase/auth/getRedirectPath";
 
-export async function signIn(formData) {
-    const email = formData.get('email');
-    const password = formData.get('password');
-
+export async function signIn({email, password}) {
     if (!email || !password) {
         throw new Error("Email and password are required");
     }
