+++ src/components/auth/SigninForm.jsx
@@ -3,66 +3,107 @@
 import Link from "next/link";
 import { useState } from "react";
 import { signIn } from "actions/auth/signin";
+import { useForm, useWatch } from "react-hook-form";
+import { signInSchema } from "lib/validation/schemas";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { getCurrentInputState } from "@/app/actions/validation/getCurrentInputState";
+import Field from "components/ui/Field";
 
 export default function SigninForm() {
-    const [isLoading, setIsLoading] = useState(false);
+    const {
+        register,
+        handleSubmit,
+        control,
+        formState: { errors, isSubmitting },
+    } = useForm({
+        resolver: zodResolver(signInSchema),
+        defaultValues: {
+            email: "",
+            password: "",
+        },
+        mode: "onBlur",
+    });
 
-    const handleFormSubmit = () => {
-        setIsLoading(true);
+    const [focusFields, setFocusFields] = useState({
+        email: false,
+        password: false,
+    });
+
+    const emailValue = useWatch({ control, name: "email" }) || "";
+    const passwordValue = useWatch({ control, name: "password" }) || "";
+
+    const onSubmit = async (data) => {
+        try {
+            await signIn(data);
+        } catch (error) {
+            console.error("Sign-in error:", error);
+        }
     };
+
+    const setFocus = (field, isFocused) => {
+        setFocusFields((prev) => ({ ...prev, [field]: isFocused }));
+    };
+
+    const emailInfo = getCurrentInputState(
+        "email",
+        emailValue,
+        errors.email,
+        focusFields.email,
+    );
+    const passwordInfo = getCurrentInputState(
+        "password",
+        passwordValue,
+        errors.password,
+        focusFields.password,
+    );
 
     return (
         <form
             className="flex flex-col gap-6 w-full"
-            action={signIn}
-            onSubmit={handleFormSubmit}
-        >
+            onSubmit={handleSubmit(onSubmit)}
+        >   
             <section className="flex flex-col gap-4 sm:gap-5 w-full">
-                <div>
-                    <label htmlFor="email" className="block text-sm font-medium text-gray-700 mb-1">
-                        Email
-                    </label>
-                    <input
-                        id="email"
-                        name="email"
-                        type="email"
-                        required
-                        className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-green-500 focus:border-transparent"
-                        placeholder="Enter your email"
-                        disabled={isLoading}
-                    />
-                </div>
 
-                <div>
-                    <label htmlFor="password" className="block text-sm font-medium text-gray-700 mb-1">
-                        Password
-                    </label>
-                    <input
-                        id="password"
-                        name="password"
+                <Field
+                    type="email"
+                    placeholder="Enter your email"
+                    registration={register("email")}
+                    state={emailInfo.state}
+                    message={emailInfo.message}
+                    disabled={isSubmitting}
+                    onFocus={() => setFocus("email", true)}
+                    onBlur={() => setFocus("email", false)}
+                />
+
+                <div className="flex flex-col gap-2 w-full">
+                    <Field
                         type="password"
-                        required
-                        className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-green-500 focus:border-transparent"
                         placeholder="Enter your password"
-                        disabled={isLoading}
+                        registration={register("password")}
+                        state={passwordInfo.state}
+                        message={passwordInfo.message}
+                        disabled={isSubmitting}
+                        onFocus={() => setFocus("password", true)}
+                        onBlur={() => setFocus("password", false)}
                     />
-                    <div className="flex justify-end w-full px-1 mt-1">
+                    <div className="flex justify-end w-full px-1">
                         <Link
                             href="/reset-password"
-                            className="text-primary text-sm hover:underline hover:underline-offset-2 transition-all duration-200"
+                            className="text-primary text-sm sm:text-base hover:underline hover:underline-offset-2 transition-all duration-200"
                         >
                             Forgot password?
                         </Link>
                     </div>
                 </div>
-            </section>
 
+            </section>
+            
             <button
                 type="submit"
                 className="btn btn-primary btn-md sm:btn-lg w-full mt-2"
-                disabled={isLoading}
+                disabled={isSubmitting}
             >
-                {isLoading ? (
+                {isSubmitting ? (
                     <>
                         <span className="loading loading-spinner"></span>
                         Signing in...
