+++ src/components/dashboard/MeSidebar.jsx
@@ -3,11 +3,10 @@
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import SidebarUserProfile from "./SidebarUserProfile";
-import { DashboardIcon, CheckboxIcon } from "@radix-ui/react-icons";
+import { DashboardIcon } from "@radix-ui/react-icons";
 
 const NAV_LINKS = [
-    { label: "Overview", href: "overview", icon: DashboardIcon },
-    { label: "Pending Approvals", href: "pending-approvals", icon: CheckboxIcon },
+    { label: "Overview", href: "overview", icon: DashboardIcon }
 ];
 
 export default function MeSidebar({ user }) {
