+++ src/proxy.js
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 
-export default async function proxy(request) {
+export async function proxy(request) {
     const isPrefetch =
         request.headers.get("next-router-prefetch") === "1" ||
         request.headers.get("purpose") === "prefetch";
